requirements.txt
# pip install -r requirements.txt

fire
bottle==0.12.25
cheroot
# gunicorn==20.1.0
requests
chardet


psutil

gevent-websocket
py-cpuinfo==9.0.0